codecheck/agents/requirements.txt
anthropic==0.18.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1